app/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bleach==6.3.0
blinker==1.9.0
cachelib==0.13.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pyotp==2.9.0
python-magic==0.4.27
qrcode==8.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.5
